+++ R/is_daylight.R
@@ -10,7 +10,7 @@
 #' that timezone correctly set in the object.
 #' @export
 is_daylight <- function(dt, lat, lon, tz) {
-  utc <- convert_to_utc(dt, tz)
+  utc <- apply_timezone(dt, tz)
   daytime <- rep(FALSE, length(dt))
 
   # All the dates in the interval